Kruskals works.  Must use only numbers for vertices in a txt file.  I replaced int with double in EdgeHeapNode constructor to make it work.

+++ trunk/ kruskals-alg/KruskalsAlgorithm/src/EdgeHeapNode.java
@@ -31,7 +31,7 @@ public class EdgeHeapNode implements Comparable {
      * @param v2	the second vertex that makes up the edge
      *                          
      */
-    public EdgeHeapNode (int key, Vertex v1, Vertex v2) {
+    public EdgeHeapNode (double key, Vertex v1, Vertex v2) {
         this.weight = key;
         this.v1 = v1;
         this.v2 = v2;
